move addEnumSuffixToModuleTitles function to utils
