+++ assembl/static2/js/app/components/administration/landingPage/manageModules.jsx
@@ -4,8 +4,6 @@ import { connect } from 'react-redux';
 import { I18n, Translate } from 'react-redux-i18n';
 import type { List, Map } from 'immutable';
 import { Button } from 'react-bootstrap';
-import countBy from 'lodash/countBy';
-import get from 'lodash/get';
 import { displayModal, closeModal } from '../../../utils/utilityManager';
 import ModulesPreview from './modulesPreview';
 import SectionTitle from '../../administration/sectionTitle';
@@ -28,7 +26,7 @@ export type LandingPageModuleType = {
   title: string
 };
 
-type LandingPageModule = {
+export type LandingPageModule = {
   configuration: Object,
   enabled: boolean,
   existsInDatabase: true,
@@ -64,22 +62,6 @@ const MODULES_IDENTIFIERS = {
   partners: 'PARTNERS',
   introduction: 'INTRODUCTION',
   footer: 'FOOTER'
-};
-
-export const addEnumSuffixToModuleTitles = (modules: Array<LandingPageModule>): Array<LandingPageModule> => {
-  // Add a suffix to the title of the module if this module appears more than one time.
-  // This suffix is the number of times this module appeared in the array.
-  const moduleTypeTitles = modules.map(module => ({ title: module.moduleType.title }));
-  const titleCounts = countBy(moduleTypeTitles, 'title');
-  const duplicatesCurrentIndex = {};
-  return modules.map((module) => {
-    const { title } = module.moduleType;
-    if (title && titleCounts[title] > 1) {
-      duplicatesCurrentIndex[title] = get(duplicatesCurrentIndex, title, 0) + 1;
-      return { ...module, moduleType: { ...module.moduleType, title: title && `${title} ${duplicatesCurrentIndex[title]}` } };
-    }
-    return module;
-  });
 };
 
 export const sortByTitle = (modules: Array<LandingPageModuleType>): Array<LandingPageModuleType> => {
