+++ assembl/static2/js/app/components/form/utils.js
@@ -1,7 +1,10 @@
 // @flow
 import { I18n } from 'react-redux-i18n';
+import countBy from 'lodash/countBy';
+import get from 'lodash/get';
 
 import type { I18nValue, FileValue, FileVariable, MutationsPromises, I18nRichTextValue, SaveStatus } from './types.flow';
+import type LandingPageModule from '../administration/landingPage/manageModules';
 import { convertEditorStateToHTML, convertEntriesToEditorState } from '../../utils/draftjs';
 import { displayAlert } from '../../utils/utilityManager';
 import { runSerial } from '../administration/saveButton';
@@ -88,3 +91,19 @@ export function getFileVariable(img: FileValue, initialImg: ?FileValue): FileVar
   const variab = img && img.externalUrl instanceof File ? img.externalUrl : null;
   return variab;
 }
+
+export const addEnumSuffixToModuleTitles = (modules: Array<LandingPageModule>): Array<LandingPageModule> => {
+  // Add a suffix to the title of the module if this module appears more than one time.
+  // This suffix is the number of times this module appeared in the array.
+  const moduleTypeTitles = modules.map(module => ({ title: module.moduleType.title }));
+  const titleCounts = countBy(moduleTypeTitles, 'title');
+  const duplicatesCurrentIndex = {};
+  return modules.map((module) => {
+    const { title } = module.moduleType;
+    if (title && titleCounts[title] > 1) {
+      duplicatesCurrentIndex[title] = get(duplicatesCurrentIndex, title, 0) + 1;
+      return { ...module, moduleType: { ...module.moduleType, title: title && `${title} ${duplicatesCurrentIndex[title]}` } };
+    }
+    return module;
+  });
+};
