+++ assembl/static2/tests/unit/components/administration/landingPage/manageModules.spec.jsx
@@ -1,11 +1,7 @@
 import React from 'react';
 import ShallowRenderer from 'react-test-renderer/shallow';
 
-import {
-  DumbManageModules,
-  addEnumSuffixToModuleTitles,
-  sortByTitle
-} from '../../../../../js/app/components/administration/landingPage/manageModules';
+import { DumbManageModules, sortByTitle } from '../../../../../js/app/components/administration/landingPage/manageModules';
 import { enabledModules, moduleTypes, modulesById } from './fakeData';
 
 describe('ManageModules component', () => {
@@ -29,57 +25,6 @@ describe('ManageModules component', () => {
     shallowRenderer.render(<DumbManageModules {...props} />);
     const result = shallowRenderer.getRenderOutput();
     expect(result).toMatchSnapshot();
-  });
-});
-
-describe('addEnumSuffixToModuleTitles function', () => {
-  it('should return the same array if there is no duplicate', () => {
-    const fakeModules = [
-      { moduleType: { title: 'lorem' } },
-      { moduleType: { title: 'ipsum' } },
-      { moduleType: { title: 'foo' } }
-    ];
-    expect(addEnumSuffixToModuleTitles(fakeModules)).toEqual(fakeModules);
-  });
-  it('should return an array of modules with count added on the title of duplicates', () => {
-    const fakeModules = [
-      { moduleType: { title: 'lorem' } },
-      { moduleType: { title: 'lorem' } },
-      { moduleType: { title: 'ipsum' } },
-      { moduleType: { title: 'foo' } }
-    ];
-    const actual = addEnumSuffixToModuleTitles(fakeModules);
-    const expected = [
-      { moduleType: { title: 'lorem 1' } },
-      { moduleType: { title: 'lorem 2' } },
-      { moduleType: { title: 'ipsum' } },
-      { moduleType: { title: 'foo' } }
-    ];
-    expect(actual).toEqual(expected);
-  });
-  it('should return an array of modules with count added on the title of different duplicates', () => {
-    const fakeModules = [
-      { moduleType: { title: 'ipsum' } },
-      { moduleType: { title: 'lorem' } },
-      { moduleType: { title: 'lorem' } },
-      { moduleType: { title: 'lorem' } },
-      { moduleType: { title: 'ipsum' } },
-      {
-        moduleType: { title: 'foo' }
-      }
-    ];
-    const actual = addEnumSuffixToModuleTitles(fakeModules);
-    const expected = [
-      { moduleType: { title: 'ipsum 1' } },
-      { moduleType: { title: 'lorem 1' } },
-      { moduleType: { title: 'lorem 2' } },
-      { moduleType: { title: 'lorem 3' } },
-      { moduleType: { title: 'ipsum 2' } },
-      {
-        moduleType: { title: 'foo' }
-      }
-    ];
-    expect(actual).toEqual(expected);
   });
 });
